main-cli: stop setting argv1
Prior to this commit, there was a global variable argv1 which is set to
the filename passed to the program. This is used by the GUI in order to
implement the much-maligned "basic mode", see
https://code.google.com/archive/p/rawtherapee/issues/2254

However, it is not used by the CLI or by options.cc at all. So we can
remove it entirely from the CLI, add a couple `extern` declarations in
the two places in the GUI where it's used, and simplify the logic.

Note that this preserves the user-unfriendly behavior where if the user
provides a filename without `-c`, presumably imagining this to be an
input filename, the tool will silently do nothing and return 1. We
should improve this to output an explicit warning or error, but we
first need to refactor the argument parsing..