main-cli: move directory listing etc out of argument parsing
In the next commit we want to split off all the pure argument parsing
from the stuff that needs access to the options, filesystem, etc etc.
To do this, we need to pull a big blob of directory-traversal logic
out of the parsing switch statement.

This also reduces the amount of greatly-nested code and makes things
easier to follow.